feat: Add Analytics class for visual data representation with multiple chart types
- Implemented the Analytics class extending BaseSolution for generating line, bar, pie, and area charts.
- Added methods for processing image data and updating charts based on object detection and tracking.
- Integrated matplotlib for chart rendering and visualization.
- Included functionality for class-wise object counting and total counts tracking.
- Enhanced user experience with customizable chart attributes and cyclic color mapping.
